Delete ws guard
